pyproject.toml
[tool.poetry]
name = "pomoberry"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {version = "0.115.0.dev.1655918067", extras = ["channels"]}

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
